Reorganize to util to match master.
