+++ indra_db/util/__init__.py
@@ -5,740 +5,14 @@ Some key functions' capabilities include:
 - inserting statements, which are stored in multiple tables, into the database.
 - distilling and deleting statements
 """
-from collections import defaultdict
 
-__all__ = ['get_defaults', 'get_primary_db', 'get_db', 'insert_raw_agents',
-           'insert_pa_stmts', 'insert_db_stmts', 'get_raw_stmts_frm_db_list',
-           'distill_stmts', 'regularize_agent_id', 'unpack']
-
-import re
-import json
-import zlib
-import logging
-
-from indra.util.nested_dict import NestedDict
-from indra.util.get_version import get_version
-from indra.util import clockit
-from indra.statements import Complex, SelfModification, ActiveForm,\
-    Conversion, Translocation, Statement
-
-from indra_db.managers.database_manager import DatabaseManager
-from indra_db.exceptions import IndraDbException
-from indra_db.config import get_databases as get_defaults
-
-logger = logging.getLogger('db_util')
-
-
-__PRIMARY_DB = None
-
-
-def get_primary_db(force_new=False):
-    """Get a DatabaseManager instance for the primary database host.
-
-    The primary database host is defined in the defaults.txt file, or in a file
-    given by the environment variable DEFAULTS_FILE. Alternatively, it may be
-    defined by the INDRADBPRIMARY environment variable. If none of the above
-    are specified, this function will raise an exception.
-
-    Note: by default, calling this function twice will return the same
-    `DatabaseManager` instance. In other words::
-
-        db1 = get_primary_db()
-        db2 = get_primary_db()
-        db1 is db2
-
-    This means also that, for example `db1.select_one(db2.TextRef)` will work,
-    in the above context.
-
-    It is still recommended that when creating a script or function, or other
-    general application, you should not rely on this feature to get your access
-    to the database, as it can make substituting a different database host both
-    complicated and messy. Rather, a database instance should be explicitly
-    passed between different users as is done in `get_statements_by_gene_role_type`
-    function's call to `get_statements` in `indra.db.query_db_stmts`.
-
-    Parameters
-    ----------
-    force_new : bool
-        If true, a new instance will be created and returned, regardless of
-        whether there is an existing instance or not. Default is False, so that
-        if this function has been called before within the global scope, a the
-        instance that was first created will be returned.
-
-    Returns
-    -------
-    primary_db : :py:class:`DatabaseManager`
-        An instance of the database manager that is attached to the primary
-        database.
-    """
-    defaults = get_defaults()
-    if 'primary' in defaults.keys():
-        primary_host = defaults['primary']
-    else:
-        raise IndraDbException("No primary host available in defaults file.")
-
-    global __PRIMARY_DB
-    if __PRIMARY_DB is None or force_new:
-        __PRIMARY_DB = DatabaseManager(primary_host, label='primary')
-        __PRIMARY_DB.grab_session()
-    return __PRIMARY_DB
-
-
-def get_test_db():
-    """Get a DatabaseManager for the test database."""
-    defaults = get_defaults()
-    test_defaults = {k: v for k, v in defaults.items() if 'test' in k}
-    key_list = list(test_defaults.keys())
-    key_list.sort()
-    db = None
-    for k in key_list:
-        test_name = test_defaults[k]
-        m = re.match('(\w+)://.*?/([\w.]+)', test_name)
-        if m is None:
-            logger.warning("Poorly formed db name: %s" % test_name)
-            continue
-        sqltype = m.groups()[0]
-        try:
-            db = DatabaseManager(test_name, sqltype=sqltype, label=k)
-            db.grab_session()
-        except Exception as e:
-            logger.error("%s didn't work" % test_name)
-            logger.exception(e)
-            continue  # Clearly this test database won't work.
-        logger.info("Using test database %s." % k)
-        break
-    if db is None:
-        logger.error("Could not find any test database names.")
-    return db
-
-
-def get_db(db_label):
-    """Get a db instance base on it's name in the config or env."""
-    defaults = get_defaults()
-    db_name = defaults[db_label]
-    m = re.match('(\w+)://.*?/([\w.]+)', db_name)
-    if m is None:
-        logger.error("Poorly formed db name: %s" % db_name)
-        return
-    sqltype = m.groups()[0]
-    return DatabaseManager(db_name, sqltype=sqltype, label=db_label)
-
-
-@clockit
-def insert_raw_agents(db, batch_id, verbose=False, num_per_yield=100):
-    """Insert agents for statements that don't have any agents.
-
-    Parameters
-    ----------
-    db : :py:class:`DatabaseManager`
-        The manager for the database into which you are adding agents.
-    batch_id : int
-        Every set of new raw statements must be given an id unique to that copy
-        That id is used to get the set of statements that need agents added.
-    verbose : bool
-        If True, print extra information and a status bar while compiling
-        agents for insert from statements. Default False.
-    num_per_yield : int
-        To conserve memory, statements are loaded in batches of `num_per_yeild`
-        using the `yeild_per` feature of sqlalchemy queries.
-    """
-    ref_tuples = []
-    mod_tuples = []
-    mut_tuples = []
-    q = db.filter_query([db.RawStatements.id, db.RawStatements],
-                         db.RawStatements.batch_id == batch_id)
-    if verbose:
-        num_stmts = q.count()
-        print("Loading:", end='', flush=True)
-
-    db_stmts = q.yield_per(num_per_yield)
-
-    for i, (stmt_id, db_stmt) in enumerate(db_stmts):
-        stmt = _get_statement_object(db_stmt)
-        ref_data, mod_data, mut_data = _extract_agent_data(stmt, stmt_id)
-        ref_tuples.extend(ref_data)
-        mod_tuples.extend(mod_data)
-        mut_tuples.extend(mut_data)
-
-        # Optionally print another tick on the progress bar.
-        if verbose and num_stmts > 25 and i % (num_stmts//25) == 0:
-            print('|', end='', flush=True)
-
-    if verbose and num_stmts > 25:
-        print()
-
-    db.copy('raw_agents', ref_tuples, ('stmt_id', 'db_name', 'db_id', 'role'))
-    db.copy('raw_mods', mod_tuples, ('stmt_id', 'type', 'position', 'residue',
-                                     'modified'))
-    db.copy('raw_muts', mut_tuples, ('stmt_id', 'position', 'residue_from',
-                                     'residue_to'))
-    return
-
-
-def insert_pa_agents(db, stmts, verbose=False, skip=None):
-    if skip is None:
-        skip = []
-
-    if verbose:
-        num_stmts = len(stmts)
-
-    # Construct the agent records
-    logger.info("Building data from agents for insert into pa_agents, "
-                "pa_mods, and pa_muts...")
-
-    if verbose:
-        print("Loading:", end='', flush=True)
-
-    ref_data = []
-    mod_data = []
-    mut_data = []
-    for i, stmt in enumerate(stmts):
-        refs, mods, muts = _extract_agent_data(stmt, stmt.get_hash())
-        ref_data.extend(refs)
-        mod_data.extend(mods)
-        mut_data.extend(muts)
-
-        # Optionally print another tick on the progress bar.
-        if verbose and num_stmts > 25 and i % (num_stmts//25) == 0:
-            print('|', end='', flush=True)
-
-    if verbose and num_stmts > 25:
-        print()
-
-    if 'agents' not in skip:
-        db.copy('pa_agents', ref_data,
-                ('stmt_mk_hash', 'db_name', 'db_id', 'role'), lazy=True)
-    if 'mods' not in skip:
-        db.copy('pa_mods', mod_data, ('stmt_mk_hash', 'type', 'position',
-                                      'residue', 'modified'))
-    if 'muts' not in skip:
-        db.copy('pa_muts', mut_data, ('stmt_mk_hash', 'position',
-                                      'residue_from', 'residue_to'))
-    return
-
-
-def regularize_agent_id(id_val, id_ns):
-    """Change agent ids for better search-ability and index-ability."""
-    ns_abbrevs = [('CHEBI', ':'), ('GO', ':'), ('HMDB', ''), ('PF', ''),
-                  ('IP', '')]
-    for ns, div in ns_abbrevs:
-        if id_ns.upper() == ns and id_val.startswith(ns):
-            new_id_val = id_val[len(ns) + len(div)]
-            break
-    else:
-        return id_val
-
-    # logger.info("Fixed agent id: %s -> %s" % (id_val, new_id_val))
-    return new_id_val
-
-
-def _extract_agent_data(stmt, stmt_id):
-    """Create the tuples for copying agents into the database."""
-    # Figure out how the agents are structured and assign roles.
-    ag_list = stmt.agent_list()
-    nary_stmt_types = [Complex, SelfModification, ActiveForm, Conversion,
-                       Translocation]
-    if any([isinstance(stmt, tp) for tp in nary_stmt_types]):
-        agents = {('OTHER', ag) for ag in ag_list}
-    elif len(ag_list) == 2:
-        agents = {('SUBJECT', ag_list[0]), ('OBJECT', ag_list[1])}
-    else:
-        raise IndraDbException("Unhandled agent structure for stmt %s "
-                               "with agents: %s."
-                               % (str(stmt), str(stmt.agent_list())))
-
-    def all_agent_refs(ag):
-        """Smooth out the iteration over agents and their refs."""
-        for ns, ag_id in ag.db_refs.items():
-            if isinstance(ag_id, list):
-                for sub_id in ag_id:
-                    yield ns, sub_id
-            else:
-                yield ns, ag_id
-        yield 'NAME', ag.name
-
-    # Prep the agents for copy into the database.
-    ref_data = []
-    mod_data = []
-    mut_data = []
-    for role, ag in agents:
-        # If no agent, or no db_refs for the agent, skip the insert
-        # that follows.
-        if ag is None or ag.db_refs is None:
-            continue
-
-        # Get the db refs data.
-        for ns, ag_id in all_agent_refs(ag):
-            if ag_id is not None:
-                ref_data.append((stmt_id, ns, regularize_agent_id(ag_id, ns),
-                                 role))
-            else:
-                logger.warning("Found agent for %s with None value." % ns)
-
-        # Get the modification data
-        for mod in ag.mods:
-            mod_data.append((stmt_id, mod.mod_type, mod.position, mod.residue,
-                             mod.is_modified))
-
-        # Get the mutation data
-        for mut in ag.mutations:
-            mut_data.append((stmt_id, mut.position, mut.residue_from,
-                             mut.residue_to))
-
-    return ref_data, mod_data, mut_data
-
-
-def insert_db_stmts(db, stmts, db_ref_id, verbose=False):
-    """Insert statement, their database, and any affiliated agents.
-
-    Note that this method is for uploading statements that came from a
-    database to our databse, not for inserting any statements to the database.
-
-    Parameters
-    ----------
-    db : :py:class:`DatabaseManager`
-        The manager for the database into which you are loading statements.
-    stmts : list [:py:class:`indra.statements.Statement`]
-        A list of un-assembled indra statements to be uploaded to the datbase.
-    db_ref_id : int
-        The id to the db_ref entry corresponding to these statements.
-    verbose : bool
-        If True, print extra information and a status bar while compiling
-        statements for insert. Default False.
-    """
-    # Preparing the statements for copying
-    stmt_data = []
-    batch_id = db.make_copy_batch_id()
-
-    cols = ('uuid', 'mk_hash', 'source_hash', 'db_info_id', 'type', 'json',
-            'indra_version', 'batch_id')
-    if verbose:
-        print("Loading:", end='', flush=True)
-    for i, stmt in enumerate(stmts):
-        # Only one evidence is allowed for each statement.
-        for ev in stmt.evidence:
-            new_stmt = stmt.make_generic_copy()
-            new_stmt.evidence.append(ev)
-            stmt_rec = (
-                new_stmt.uuid,
-                new_stmt.get_hash(shallow=False),
-                new_stmt.evidence[0].get_source_hash(),
-                db_ref_id,
-                new_stmt.__class__.__name__,
-                json.dumps(new_stmt.to_json()).encode('utf8'),
-                get_version(),
-                batch_id
-            )
-            stmt_data.append(stmt_rec)
-        if verbose and i % (len(stmts)//25) == 0:
-            print('|', end='', flush=True)
-    if verbose:
-        print(" Done loading %d statements." % len(stmts))
-    db.copy('raw_statements', stmt_data, cols, lazy=True, push_conflict=True)
-    insert_raw_agents(db, batch_id)
-    return
-
-
-def insert_pa_stmts(db, stmts, verbose=False, do_copy=True,
-                    ignore_agents=False):
-    """Insert pre-assembled statements, and any affiliated agents.
-
-    Parameters
-    ----------
-    db : :py:class:`DatabaseManager`
-        The manager for the database into which you are loading pre-assembled
-        statements.
-    stmts : iterable [:py:class:`indra.statements.Statement`]
-        A list of pre-assembled indra statements to be uploaded to the datbase.
-    verbose : bool
-        If True, print extra information and a status bar while compiling
-        statements for insert. Default False.
-    do_copy : bool
-        If True (default), use pgcopy to quickly insert the agents.
-    """
-    logger.info("Beginning to insert pre-assembled statements.")
-    stmt_data = []
-    indra_version = get_version()
-    cols = ('uuid', 'matches_key', 'mk_hash', 'type', 'json', 'indra_version')
-    if verbose:
-        print("Loading:", end='', flush=True)
-    for i, stmt in enumerate(stmts):
-        stmt_rec = (
-            stmt.uuid,
-            stmt.matches_key(),
-            stmt.get_hash(shallow=True),
-            stmt.__class__.__name__,
-            json.dumps(stmt.to_json()).encode('utf8'),
-            indra_version
-        )
-        stmt_data.append(stmt_rec)
-        if verbose and i % (len(stmts)//25) == 0:
-            print('|', end='', flush=True)
-    if verbose:
-        print(" Done loading %d statements." % len(stmts))
-    if do_copy:
-        db.copy('pa_statements', stmt_data, cols)
-    else:
-        db.insert_many('pa_statements', stmt_data, cols=cols)
-    if not ignore_agents:
-        insert_pa_agents(db, stmts, verbose=verbose)
-    return
-
-
-#==============================================================================
-# Below are some functions that are useful for getting raw statements from the
-# database at various levels of abstraction.
-#==============================================================================
-
-def _get_statement_object(db_stmt):
-    """Get an INDRA Statement object from a db_stmt."""
-    return Statement._from_json(json.loads(db_stmt.json.decode('utf-8')))
-
-
-def _set_evidence_text_ref(stmt, tr):
-    # This is a separate function because it is likely to change, and this is a
-    # critical process that is executed in multiple places.
-    for ev in stmt.evidence:
-        ev.pmid = tr.pmid
-
-
-@clockit
-def _fix_evidence_refs(db, rid_stmt_trios):
-    """Get proper id data for a raw statement from the database.
-
-    Alterations are made to the Statement objects "in-place", so this function
-    itself returns None.
-    """
-    rid_set = {rid for rid, _, _ in rid_stmt_trios if rid is not None}
-    logger.info("Getting text refs for %d readings." % len(rid_set))
-    if rid_set:
-        rid_tr_pairs = db.select_all(
-            [db.Reading.id, db.TextRef],
-            db.Reading.id.in_(rid_set),
-            db.Reading.text_content_id == db.TextContent.id,
-            db.TextContent.text_ref_id == db.TextRef.id
-            )
-        rid_tr_dict = {rid: tr for rid, tr in rid_tr_pairs}
-        for rid, sid, stmt in rid_stmt_trios:
-            if rid is None:
-                # This means this statement came from a database, not reading.
-                continue
-            assert len(stmt.evidence) == 1, \
-                "Only raw statements can have their refs fixed."
-            _set_evidence_text_ref(stmt, rid_tr_dict[rid])
-    return
-
-
-@clockit
-def get_raw_stmts_frm_db_list(db, db_stmt_objs, fix_refs=True, with_sids=True):
-    """Convert table objects of raw statements into INDRA Statement objects."""
-    rid_stmt_sid_trios = [(db_stmt.reading_id, db_stmt.id,
-                          _get_statement_object(db_stmt))
-                          for db_stmt in db_stmt_objs]
-    if fix_refs:
-        _fix_evidence_refs(db, rid_stmt_sid_trios)
-    # Note: it is important that order is maintained here (hence not a set or
-    # dict).
-    if with_sids:
-        return [(sid, stmt) for _, sid, stmt in rid_stmt_sid_trios]
-    else:
-        return [stmt for _, _, stmt in rid_stmt_sid_trios]
-
-
-def _get_reading_statement_dict(db, clauses=None, get_full_stmts=True):
-    """Get a nested dict of statements, keyed by ref, content, and reading."""
-    # Construct the query for metadata from the database.
-    q = (db.session.query(db.TextRef, db.TextContent.id,
-                          db.TextContent.source, db.Reading.id,
-                          db.Reading.reader_version, db.RawStatements.id,
-                          db.RawStatements.json)
-         .filter(db.RawStatements.reading_id == db.Reading.id,
-                 db.Reading.text_content_id == db.TextContent.id,
-                 db.TextContent.text_ref_id == db.TextRef.id))
-    if clauses:
-        q = q.filter(*clauses)
-
-    # Prime some counters.
-    num_duplicate_evidence = 0
-    num_unique_evidence = 0
-
-    # Populate a dict with all the data.
-    stmt_nd = NestedDict()
-    for tr, tcid, src, rid, rv, sid, sjson in q.yield_per(1000):
-        # Back out the reader name.
-        for reader, rv_list in reader_versions.items():
-            if rv in rv_list:
-                break
-        else:
-            raise Exception("rv %s not recognized." % rv)
-
-        # Get the json for comparison and/or storage
-        stmt_json = json.loads(sjson.decode('utf8'))
-        stmt = Statement._from_json(stmt_json)
-        _set_evidence_text_ref(stmt, tr)
-
-        # Hash the compbined stmt and evidence matches key.
-        stmt_hash = stmt.get_hash(shallow=False)
-
-        # For convenience get the endpoint statement dict
-        s_dict = stmt_nd[tr.id][src][tcid][reader][rv][rid]
-
-        # Initialize the value to a set, and count duplicates
-        if stmt_hash not in s_dict.keys():
-            s_dict[stmt_hash] = set()
-            num_unique_evidence += 1
-        else:
-            num_duplicate_evidence += 1
-
-        # Either store the statement, or the statement id.
-        if get_full_stmts:
-            s_dict[stmt_hash].add((sid, stmt))
-        else:
-            s_dict[stmt_hash].add((sid, None))
-
-    # Report on the results.
-    print("Found %d relevant text refs with statements." % len(stmt_nd))
-    print("number of statement exact duplicates: %d" % num_duplicate_evidence)
-    print("number of unique statements: %d" % num_unique_evidence)
-    return stmt_nd
-
-
-# Specify versions of readers, and preference. Later in the list is better.
-reader_versions = {
-    'sparser': ['sept14-linux\n', 'sept14-linux', 'June2018-linux',
-                'October2018-linux'],
-    'reach': ['61059a-biores-e9ee36', '1.3.3-61059a-biores-']
-    }
-
-# Specify sources of fulltext content, and order priorities.
-text_content_sources = ['pubmed', 'elsevier', 'manuscripts', 'pmc_oa']
-
-
-def _get_filtered_rdg_statements(stmt_nd, get_full_stmts, linked_sids=None,
-                                 ignore_duplicates=False):
-    """Get the set of statements/ids from readings minus exact duplicates."""
-    logger.info("Filtering the statements from reading.")
-    if linked_sids is None:
-        linked_sids = set()
-
-    # Now we filter and get the set of statements/statement ids.
-    stmt_tpls = set()
-    bettered_duplicate_sids = set()  # Statements with "better" alternatives
-    for trid, src_dict in stmt_nd.items():
-        some_bettered_duplicate_tpls = set()
-
-        # Filter out the older reader versions
-        for reader, rv_list in reader_versions.items():
-            simple_src_dict = defaultdict(dict)
-            for (src, _, _), rv_dict in src_dict.get_paths(reader):
-                best_rv = max(rv_dict, key=lambda x: rv_list.index(x))
-
-                # Record the rest of the statement ids.
-                for rv, r_dict in rv_dict.items():
-                    if rv != best_rv:
-                        some_bettered_duplicate_tpls |= r_dict.get_leaves()
-                    else:
-                        for h, stmt_set in list(r_dict.values())[0].items():
-                            # Sort the statements by their source, and whether
-                            # they are new or old, defined as whether they
-                            # have yet been included in preassembly. There
-                            # should be no overlap in hashes here.
-                            sid_set = {sid for sid, _ in stmt_set}
-                            if sid_set < linked_sids:
-                                simple_src_dict[src][h] = ('old', stmt_set)
-                            elif sid_set.isdisjoint(linked_sids):
-                                simple_src_dict[src][h] = ('new', stmt_set)
-                            else:
-                                # If you ever see this pop up, something very
-                                # strange has happened. It means that some
-                                # statements from a given reading were already
-                                # included in preassembly, but others were not.
-                                # There is no mechanism that should accept only
-                                # some statements from within a reading, so
-                                # that should never happen, but Murphy will
-                                # always win the day.
-                                assert False, \
-                                    "Found reading partially included."
-
-            # Choose the statements to propagate
-            new_stmt_dict = {}
-            for src in reversed(text_content_sources):
-                for h, (status, s_set) in simple_src_dict[src].items():
-                    # If this error ever comes up, it means that the uniqueness
-                    # constraint on raw statements per reading is not
-                    # functioning correctly.
-                    assert len(s_set) == 1, \
-                        "Found exact duplicates from the same reading."
-
-                    # Choose whether to keep the statement or not.
-                    s_tpl = s_set.pop()
-                    if h not in new_stmt_dict:
-                        # No conflict, no problem
-                        new_stmt_dict[h] = s_tpl
-                    elif status == 'old':
-                        # The same statement was newly found by a better
-                        # version.
-                        some_bettered_duplicate_tpls.add(s_tpl)
-            stmt_tpls |= set(new_stmt_dict.values())
-
-        # Add the bettered duplicates found in this round.
-        bettered_duplicate_sids |= \
-            {sid for sid, _ in some_bettered_duplicate_tpls}
-
-    if get_full_stmts:
-        stmts = {stmt for _, stmt in stmt_tpls if stmt is not None}
-        assert len(stmts) == len(stmt_tpls),\
-            ("Some statements were None! The interaction between "
-             "_get_reading_statement_dict and _filter_rdg_statements was "
-             "probably mishandled.")
-    else:
-        stmts = {sid for sid, _ in stmt_tpls}
-
-    return stmts, bettered_duplicate_sids
-
-
-def _get_filtered_db_statements(db, get_full_stmts=False, clauses=None):
-    """Get the set of statements/ids from databases minus exact duplicates."""
-    # Only get the json if it's going to be used.
-    if get_full_stmts:
-        tbl_list = [db.RawStatements.json]
-    else:
-        tbl_list = [db.RawStatements.id]
-
-    db_s_q = db.filter_query(tbl_list, db.RawStatements.db_info_id.isnot(None))
-
-    # Add any other criterion specified at higher levels.
-    if clauses:
-        db_s_q = db_s_q.filter(*clauses)
-
-    # Produce a generator of statement groups.
-    db_stmt_data = db_s_q.yield_per(10000)
-    if get_full_stmts:
-        return {Statement._from_json(json.loads(s_json.decode('utf-8')))
-                for s_json, in db_stmt_data}
-    else:
-        return {sid for sid, in db_stmt_data}
-
-
-@clockit
-def distill_stmts(db, get_full_stmts=False, clauses=None,
-                  handle_duplicates='ignore', weed_evidence=True):
-    """Get a corpus of statements from clauses and filters duplicate evidence.
-
-    Parameters
-    ----------
-    db : :py:class:`DatabaseManager`
-        A database manager instance to access the database.
-    get_full_stmts : bool
-        By default (False), only Statement ids (the primary index of Statements
-        on the database) are returned. However, if set to True, serialized
-        INDRA Statements will be returned. Note that this will in general be
-        VERY large in memory, and therefore should be used with caution.
-    clauses : None or list of sqlalchemy clauses
-        By default None. Specify sqlalchemy clauses to reduce the scope of
-        statements, e.g. `clauses=[db.Statements.type == 'Phosphorylation']` or
-        `clauses=[db.Statements.uuid.in_([<uuids>])]`.
-    handle_duplicates : 'ignore', 'delete', or a string file path
-        Choose whether you want to delete the statements that are found to be
-        duplicates ('delete'), or write a pickle file with their ids (at the
-        string file path) for later handling, or simply do nothing ('ignore').
-        The default behavior is 'ignore'.
-    weed_evidence : bool
-        If True, evidence links that exist for raw statements that now have
-        better alternatives will be removed. If False, such links will remain,
-        which may cause problems in incremental pre-assembly.
-
-    Returns
-    -------
-    stmt_ret : set
-        A set of either statement ids or serialized statements, depending on
-        `get_full_stmts`.
-    """
-    if handle_duplicates == 'delete' or handle_duplicates != 'ignore':
-        logger.info("Looking for ids from existing links...")
-        linked_sids = {sid for sid,
-                       in db.select_all(db.RawUniqueLinks.raw_stmt_id)}
-    else:
-        linked_sids = set()
-
-    # Get de-duplicated Statements, and duplicate uuids, as well as uuid of
-    # Statements that have been improved upon...
-    logger.info("Sorting reading statements...")
-    stmt_nd = _get_reading_statement_dict(db, clauses, get_full_stmts)
-
-    stmts, bettered_duplicate_sids = \
-        _get_filtered_rdg_statements(stmt_nd, get_full_stmts, linked_sids)
-    logger.info("After filtering reading: %d unique statements, and %d with "
-                "results from better resources available."
-                % (len(stmts), len(bettered_duplicate_sids)))
-    del stmt_nd  # This takes up a lot of memory, and is done being used.
-
-    db_stmts = _get_filtered_db_statements(db, get_full_stmts, clauses)
-    stmts |= db_stmts
-
-    # Remove support links for statements that have better versions available.
-    bad_link_sids = bettered_duplicate_sids & linked_sids
-    if len(bad_link_sids) and weed_evidence:
-        logger.info("Removing bettered evidence links...")
-        rm_links = db.select_all(
-            db.RawUniqueLinks,
-            db.RawUniqueLinks.raw_stmt_id.in_(bad_link_sids)
-            )
-        db.delete_all(rm_links)
-
-    return stmts
-
-
-def delete_raw_statements_by_id(db, raw_sids, sync_session=False,
-                                remove='all'):
-    """Delete raw statements, their agents, and their raw-unique links.
-
-    It is best to batch over this function with sets of 1000 or so ids. Setting
-    sync_session to False will result in a much faster resolution, but you may
-    find some ORM objects have not been updated.
-    """
-    if remove == 'all':
-        remove = ['links', 'agents', 'statements']
-
-    # First, delete the evidence links.
-    if 'links' in remove:
-        ev_q = db.filter_query(db.RawUniqueLinks,
-                               db.RawUniqueLinks.raw_stmt_id.in_(raw_sids))
-        logger.info("Deleting any connected evidence links...")
-        ev_q.delete(synchronize_session=sync_session)
-
-    # Second, delete the agents.
-    if 'agents' in remove:
-        ag_q = db.filter_query(db.RawAgents,
-                               db.RawAgents.stmt_id.in_(raw_sids))
-        logger.info("Deleting all connected agents...")
-        ag_q.delete(synchronize_session=sync_session)
-
-    # Now finally delete the statements.
-    if 'statements' in remove:
-        raw_q = db.filter_query(db.RawStatements,
-                                db.RawStatements.id.in_(raw_sids))
-        logger.info("Deleting all raw indicated statements...")
-        raw_q.delete(synchronize_session=sync_session)
-    return
-
-
-def unpack(bts, decode=True):
-    ret = zlib.decompress(bts, zlib.MAX_WBITS+16)
-    if decode:
-        ret = ret.decode('utf-8')
-    return ret
-
-
-def _get_trids(db, id_val, id_type):
-    """Return text ref IDs corresponding to any ID type and value."""
-    # Get the text ref id(s)
-    if id_type in ['trid']:
-        trids = [int(id_val)]
-    else:
-        id_types = ['pmid', 'pmcid', 'doi', 'pii', 'url', 'manuscript_id']
-        if id_type not in id_types:
-            raise ValueError('id_type must be one of: %s' % str(id_types))
-        constraint = (getattr(db.TextRef, id_type) == id_val)
-        trids = [trid for trid, in db.select_all(db.TextRef.id, constraint)]
-    return trids
+__all__ = ['get_primary_db', 'get_db', 'insert_raw_agents', 'insert_pa_stmts',
+           'insert_pa_agents', 'insert_db_stmts', 'get_raw_stmts_frm_db_list',
+           'distill_stmts', 'regularize_agent_id', 'get_statement_object']
+
+from .insert import *
+from .helpers import *
+from .dump_sif import *
+from .constructors import *
+from .content_scripts import *
+from .distill_statements import *
